Move dashboard homepage queires to homepage tests
